BookController and view updated to display all books
The BookController's index method has been updated to retrieve all books using the Book model and pass it to the view. The view 'book.book' has been updated to display the list of books. The navigation link in the sidebar has also been updated to reflect the change from "Courses" to "Books".